# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamBox — Movie & Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AnimeItem {
  mal_id: number;
  title: string;
  /** @nullable */
  title_english?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  episodes?: number | null;
  status: string;
  /** @nullable */
  year?: number | null;
  genres: string[];
  /** @nullable */
  type?: string | null;
}

export interface AnimeDetail {
  mal_id: number;
  title: string;
  /** @nullable */
  title_english?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  trailer_url?: string | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  episodes?: number | null;
  status: string;
  /** @nullable */
  year?: number | null;
  genres: string[];
  studios: string[];
  /** @nullable */
  type?: string | null;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  rating?: string | null;
}

export interface AnimeListResponse {
  items: AnimeItem[];
  total: number;
  page: number;
  hasNext: boolean;
}

export interface EpisodeItem {
  number: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  aired?: string | null;
  filler: boolean;
}

export interface EpisodeListResponse {
  items: EpisodeItem[];
  page: number;
  hasNext: boolean;
}

export type EmbedResponseSourcesItem = {
  label: string;
  url: string;
};

export interface EmbedResponse {
  embedUrl: string;
  sources: EmbedResponseSourcesItem[];
}

export interface MovieItem {
  tmdb_id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  backdrop?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  year?: number | null;
  genres: string[];
}

export interface MovieDetail {
  tmdb_id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  backdrop?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  runtime?: number | null;
  genres: string[];
  cast: string[];
  /** @nullable */
  trailer_url?: string | null;
  /** @nullable */
  imdb_id?: string | null;
}

export interface MovieListResponse {
  items: MovieItem[];
  total: number;
  page: number;
  hasNext: boolean;
}

export interface GenreItem {
  id: number;
  name: string;
}

export interface GenreListResponse {
  genres: GenreItem[];
}

export type WatchlistItemContentType = typeof WatchlistItemContentType[keyof typeof WatchlistItemContentType];


export const WatchlistItemContentType = {
  anime: 'anime',
  movie: 'movie',
} as const;

export interface WatchlistItem {
  id: number;
  contentId: number;
  contentType: WatchlistItemContentType;
  title: string;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  score?: number | null;
  addedAt: string;
}

export type WatchlistInputContentType = typeof WatchlistInputContentType[keyof typeof WatchlistInputContentType];


export const WatchlistInputContentType = {
  anime: 'anime',
  movie: 'movie',
} as const;

export interface WatchlistInput {
  contentId: number;
  contentType: WatchlistInputContentType;
  title: string;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  score?: number | null;
}

export type GetAnimeTrendingParams = {
page?: number;
limit?: number;
};

export type GetAnimeTopParams = {
page?: number;
limit?: number;
genre?: string;
};

export type SearchAnimeParams = {
q: string;
page?: number;
genre?: string;
};

export type GetAnimeEpisodesParams = {
malId: number;
page?: number;
};

export type GetAnimeEmbedParams = {
malId: number;
ep?: number;
dubbed?: boolean;
};

export type GetMoviesTrendingParams = {
page?: number;
};

export type GetMoviesTopParams = {
page?: number;
genre?: string;
};

export type SearchMoviesParams = {
q: string;
page?: number;
};