# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamBox — Movie & Anime Streaming API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /anime/trending:
    get:
      operationId: getAnimeTrending
      tags: [anime]
      summary: Get currently airing / trending anime
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of trending anime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/top:
    get:
      operationId: getAnimeTop
      tags: [anime]
      summary: Get top rated anime
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: genre
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of top anime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/search:
    get:
      operationId: searchAnime
      tags: [anime]
      summary: Search anime by title
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: genre
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
...
  /anime/{id}:
    get:
      operationId: getAnimeById
      tags: [anime]
      summary: Get anime details by MAL ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Anime detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /anime/episodes:
    get:
      operationId: getAnimeEpisodes
      tags: [anime]
      summary: Get episodes list for an anime
      parameters:
        - name: malId
          in: query
          required: true
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: Episode list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeListResponse"
  /anime/embed:
    get:
      operationId: getAnimeEmbed
      tags: [anime]
      summary: Get embed URL for anime episode
      parameters:
        - name: malId
          in: query
          required: true
          schema: { type: integer }
        - name: ep
          in: query
          schema: { type: integer, default: 1 }
        - name: dubbed
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Embed info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedResponse"

  # ── MOVIES ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]

    # ── Anime schemas ──
    AnimeItem:
      type: object
      properties:
        mal_id: { type: integer }
        title: { type: string }
        title_english: { type: ["string", "null"] }
        synopsis: { type: ["string", "null"] }
        poster: { type: ["string", "null"] }
        score: { type: ["number", "null"] }
        episodes: { type: ["integer", "null"] }
        status: { type: string }
        year: { type: ["integer", "null"] }
        genres: { type: array, items: { type: string } }
        type: { type: ["string", "null"] }
      required: [mal_id, title, status, genres]
    AnimeDetail:
      type: object
      properties:
        mal_id: { type: integer }
        title: { type: string }
        title_english: { type: ["string", "null"] }
        synopsis: { type: ["string", "null"] }
        poster: { type: ["string", "null"] }
        trailer_url: { type: ["string", "null"] }
        score: { type: ["number", "null"] }
        rank: { type: ["integer", "null"] }
        episodes: { type: ["integer", "null"] }
        status: { type: string }
        year: { type: ["integer", "null"] }
        genres: { type: array, items: { type: string } }
        studios: { type: array, items: { type: string } }
        type: { type: ["string", "null"] }
        duration: { type: ["string", "null"] }
        rating: { type: ["string", "null"] }
      required: [mal_id, title, status, genres, studios]
    AnimeListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AnimeItem" } }
        total: { type: integer }
        page: { type: integer }
        hasNext: { type: boolean }
      required: [items, total, page, hasNext]
    EpisodeItem:
      type: object
      properties:
        number: { type: integer }
        title: { type: ["string", "null"] }
        aired: { type: ["string", "null"] }
        filler: { type: boolean }
      required: [number, filler]
    EpisodeListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/EpisodeItem" } }
        page: { type: integer }
        hasNext: { type: boolean }
      required: [items, page, hasNext]
    EmbedResponse:
      type: object
      properties:
        embedUrl: { type: string }
        sources:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              url: { type: string }
            required: [label, url]
      required: [embedUrl, sources]

    # ── Movie schemas ──
...