# lib/api-zod/src/generated/types/animeDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamBox — Movie & Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface AnimeDetail {
  mal_id: number;
  title: string;
  /** @nullable */
  title_english?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  trailer_url?: string | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  episodes?: number | null;
  status: string;
  /** @nullable */
  year?: number | null;
  genres: string[];
  studios: string[];
  /** @nullable */
  type?: string | null;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  rating?: string | null;
}