# lib/api-zod/src/generated/types/animeListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamBox — Movie & Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeItem } from './animeItem';

export interface AnimeListResponse {
  items: AnimeItem[];
  total: number;
  page: number;
  hasNext: boolean;
}