# lib/api-zod/src/generated/types/embedResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamBox — Movie & Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */
import type { EmbedResponseSourcesItem } from './embedResponseSourcesItem';

export interface EmbedResponse {
  embedUrl: string;
  sources: EmbedResponseSourcesItem[];
}