# lib/api-zod/src/generated/types/episodeListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamBox — Movie & Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */
import type { EpisodeItem } from './episodeItem';

export interface EpisodeListResponse {
  items: EpisodeItem[];
  page: number;
  hasNext: boolean;
}